notebook-controller/requirements-integrity.txt
#
# This file is autogenerated by pip-compile with Python 3.8
# by the following command:
#
#    pip-compile requirements-integrity.in
#
anyio==3.7.1
    # via httpcore
asttokens==2.2.1
    # via stack-data
attrs==23.1.0
    # via jsonschema
backcall==0.2.0
    # via ipython
bcrypt==4.0.1
    # via paramiko
cachetools==5.3.1
    # via google-auth
certifi==2023.7.22
    # via
    #   httpcore
    #   httpx
    #   kubernetes
    #   requests
cffi==1.15.1
    # via
    #   cryptography
    #   pynacl
charmed-kubeflow-chisme==0.2.0
    # via -r requirements-integrity.in
charset-normalizer==3.2.0
    # via requests
cryptography==41.0.3
    # via paramiko
decorator==5.1.1
    # via
    #   ipdb
    #   ipython
deepdiff==6.2.1
    # via charmed-kubeflow-chisme
exceptiongroup==1.1.3
    # via
    #   anyio
    #   pytest
executing==1.2.0
    # via stack-data
google-auth==2.22.0
    # via kubernetes
h11==0.14.0
    # via httpcore
httpcore==0.17.3
    # via httpx
httpx==0.24.1
    # via lightkube
idna==3.4
    # via
    #   anyio
    #   httpx
    #   requests
importlib-resources==6.0.1
    # via jsonschema
iniconfig==2.0.0
    # via pytest
ipdb==0.13.13
    # via pytest-operator
ipython==8.12.2
    # via ipdb
jedi==0.19.0
    # via ipython
jinja2==3.1.2
    # via
    #   charmed-kubeflow-chisme
    #   pytest-operator
jsonschema==4.17.3
    # via serialized-data-interface
juju==2.9.44.0
    # via
    #   -r requirements-integrity.in
    #   pytest-operator
jujubundlelib==0.5.7
    # via theblues
kubernetes==27.2.0
    # via juju
lightkube==0.14.0
    # via charmed-kubeflow-chisme
lightkube-models==1.27.1.4
    # via lightkube
macaroonbakery==1.3.1
    # via
    #   juju
    #   theblues
markupsafe==2.1.3
    # via jinja2
matplotlib-inline==0.1.6
    # via ipython
mypy-extensions==1.0.0
    # via typing-inspect
oauthlib==3.2.2
    # via
    #   kubernetes
    #   requests-oauthlib
ops==2.5.1
    # via
    #   -r requirements-integrity.in
    #   charmed-kubeflow-chisme
    #   serialized-data-interface
ordered-set==4.1.0
    # via deepdiff
packaging==23.1
    # via pytest
paramiko==2.12.0
    # via juju
parso==0.8.3
    # via jedi
pexpect==4.8.0
    # via ipython
pickleshare==0.7.5
    # via ipython
pkgutil-resolve-name==1.3.10
    # via jsonschema
pluggy==1.2.0
    # via pytest
prompt-toolkit==3.0.39
    # via ipython
protobuf==3.20.3
    # via macaroonbakery
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.2
    # via stack-data
pyasn1==0.5.0
    # via
    #   juju
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.3.0
    # via google-auth
pycparser==2.21
    # via cffi
pygments==2.16.1
    # via ipython
pymacaroons==0.13.0
    # via macaroonbakery
pynacl==1.5.0
    # via
    #   macaroonbakery
    #   paramiko
    #   pymacaroons
pyrfc3339==1.1
    # via
    #   juju
    #   macaroonbakery
pyrsistent==0.19.3
    # via jsonschema
pytest==7.4.0
    # via
    #   -r requirements-integrity.in
    #   pytest-asyncio
    #   pytest-operator
pytest-asyncio==0.21.1
    # via pytest-operator
pytest-operator==0.28.0
    # via -r requirements-integrity.in
python-dateutil==2.8.2
    # via kubernetes
pytz==2023.3
    # via pyrfc3339
pyyaml==6.0.1
    # via
    #   juju
    #   jujubundlelib
    #   kubernetes
    #   lightkube
    #   ops
    #   pytest-operator
    #   serialized-data-interface
requests==2.31.0
    # via
    #   kubernetes
    #   macaroonbakery
    #   requests-oauthlib
    #   serialized-data-interface
    #   theblues
requests-oauthlib==1.3.1
    # via kubernetes
rsa==4.9
    # via google-auth
ruamel-yaml==0.17.32
    # via charmed-kubeflow-chisme
ruamel-yaml-clib==0.2.7
    # via ruamel-yaml
serialized-data-interface==0.7.0
    # via charmed-kubeflow-chisme
six==1.16.0
    # via
    #   asttokens
    #   google-auth
    #   kubernetes
    #   macaroonbakery
    #   paramiko
    #   pymacaroons
    #   python-dateutil
sniffio==1.3.0
    # via
    #   anyio
    #   httpcore
    #   httpx
stack-data==0.6.2
    # via ipython
tenacity==8.2.3
    # via charmed-kubeflow-chisme
theblues==0.5.2
    # via juju
tomli==2.0.1
    # via
    #   ipdb
    #   pytest
toposort==1.10
    # via juju
traitlets==5.9.0
    # via
    #   ipython
    #   matplotlib-inline
typing-extensions==4.7.1
    # via
    #   ipython
    #   typing-inspect
typing-inspect==0.9.0
    # via juju
urllib3==1.26.16
    # via
    #   google-auth
    #   kubernetes
    #   requests
wcwidth==0.2.6
    # via prompt-toolkit
websocket-client==1.6.2
    # via
    #   kubernetes
    #   ops
websockets==7.0
    # via juju
zipp==3.16.2
    # via importlib-resources
